requirements.txt
# 项目依赖包

# Excel处理
pandas==2.3.3
openpyxl==3.1.5

# Word文档生成
python-docx==1.2.0

# 其他
lxml==5.3.1
